WAZU says http://www.wazu.jp/gallery/Fonts_Hanunoo.html
MPH 2B Damase doesn't support the consonant-vowel ligatures necessary to render Buhid writing.

OK, so I made 'mark' lookups for combining marks and a bunch of ligatures in
an 'rlig' lookup.  The latter still not working: don't know why.

Made page to match the example of the combining forms at
http://www.omniglot.com/writing/hanunoo.htm